v016.

+++ models.py
@@ -42,16 +42,16 @@ class Order(Base):
     material = Column(String, nullable=False)
     quantity = Column(Integer, nullable=False)
     performer = Column(String, nullable=False)
-    print_width = Column(Numeric(10, 2), nullable=False)
-    print_height = Column(Numeric(10, 2), nullable=False)
-    canvas_width = Column(Numeric(10, 2), nullable=False)
-    canvas_length = Column(Numeric(10, 2), nullable=False)
+    print_width = Column(Float, nullable=False)
+    print_height = Column(Float, nullable=False)
+    canvas_width = Column(Float, nullable=False)
+    canvas_length = Column(Float, nullable=False)
     eyelets = Column(String, nullable=False)
     spike = Column(String, nullable=False)
     reinforcement = Column(String, nullable=False)
     customer = Column(String, nullable=True)
-    price_per_unit = Column(Numeric(10, 2), nullable=True)
-    total_amount = Column(Numeric(10, 2), nullable=True)
+    price_per_unit = Column(Integer, nullable=True)
+    total_amount = Column(Integer, nullable=True)
 
     result: Mapped["Result"] = relationship(back_populates="order")
 
@@ -59,24 +59,24 @@ class Result(Base):
     __tablename__ = 'results'
 
     order_id = Column(Integer, ForeignKey('orders.id'), primary_key=True)
-    total_print_area = Column(Numeric(10, 2), nullable=True)
-    total_canvas_area = Column(Numeric(10, 2), nullable=True)
-    total_paints = Column(Numeric(10, 2), nullable=True)
+    total_print_area = Column(Float, nullable=True)
+    total_canvas_area = Column(Float, nullable=True)
+    total_paints = Column(Float, nullable=True)
     total_eyelets = Column(Integer, nullable=True)
     total_spikes = Column(Integer, nullable=True)
     total_reinforcements = Column(Integer, nullable=True)
 
-    expenses_canvas = Column(Numeric(10, 2), nullable=True)
-    expenses_prints = Column(Numeric(10, 2), nullable=True)
-    expenses_eyelets = Column(Numeric(10, 2), nullable=True)
-    expenses_reinforcements = Column(Numeric(10, 2), nullable=True)
-    salary_printer = Column(Numeric(10, 2), nullable=True)
-    salary_eyelet_worker = Column(Numeric(10, 2), nullable=True)
-    salary_cutter = Column(Numeric(10, 2), nullable=True)
-    salary_welder = Column(Numeric(10, 2), nullable=True)
-    total_expenses = Column(Numeric(10, 2), nullable=True)
-    tax = Column(Numeric(10, 2), nullable=True)
-    margin = Column(Numeric(10, 2), nullable=True)
+    expenses_canvas = Column(Float, nullable=True)
+    expenses_prints = Column(Float, nullable=True)
+    expenses_eyelets = Column(Float, nullable=True)
+    expenses_reinforcements = Column(Float, nullable=True)
+    salary_printer = Column(Float, nullable=True)
+    salary_eyelet_worker = Column(Float, nullable=True)
+    salary_cutter = Column(Float, nullable=True)
+    salary_welder = Column(Float, nullable=True)
+    total_expenses = Column(Float, nullable=True)
+    tax = Column(Float, nullable=True)
+    margin = Column(Float, nullable=True)
 
     order: Mapped["Order"] = relationship(back_populates="result")
 
@@ -87,27 +87,27 @@ class Settings(Base):
     id = Column(Integer, primary_key=True)
     updated_at = Column(Date, nullable=False)
 
-    blueback_price_m2 = Column(Numeric(10, 2))  # Цена м² блюбэка
-    banner_molded_price_m2 = Column(Numeric(10, 2))  # Цена м² баннера литого 450г
-    banner_laminated_price_m2 = Column(Numeric(10, 2))  # Цена м² баннера ламинат 440г
-    mesh_price_m2 = Column(Numeric(10, 2))  # Цена м² сетки
+    blueback_price_m2 = Column(Float)  # Цена м² блюбэка
+    banner_molded_price_m2 = Column(Float)  # Цена м² баннера литого 450г
+    banner_laminated_price_m2 = Column(Float)  # Цена м² баннера ламинат 440г
+    mesh_price_m2 = Column(Float)  # Цена м² сетки
 
-    blueback_price_roll = Column(Numeric(10, 2))  # Стоимость рулона блюбэка
-    banner_molded_price_roll = Column(Numeric(10, 2))  # Стоимость рулона баннера литого 450г
-    banner_laminated_price_roll = Column(Numeric(10, 2))  # Стоимость рулона баннера ламинат 440г
-    mesh_price_roll = Column(Numeric(10, 2))  # Стоимость рулона сетки
+    blueback_price_roll = Column(Float)  # Стоимость рулона блюбэка
+    banner_molded_price_roll = Column(Float)  # Стоимость рулона баннера литого 450г
+    banner_laminated_price_roll = Column(Float)  # Стоимость рулона баннера ламинат 440г
+    mesh_price_roll = Column(Float)  # Стоимость рулона сетки
 
-    eyelet_step = Column(Numeric(10, 2))  # Шаг люверса (м)
-    eyelet_price = Column(Numeric(10, 2))  # Цена люверса
+    eyelet_step = Column(Float)  # Шаг люверса (м)
+    eyelet_price = Column(Float)  # Цена люверсач
 
-    paint_price_liter = Column(Numeric(10, 2))  # Цена литра краски
-    paint_consumption_m2 = Column(Numeric(10, 2))  # Расход краски на м²
-    print_price_m2 = Column(Numeric(10, 2))  # Цена печати за м²
+    paint_price_liter = Column(Float)  # Цена литра краски
+    paint_consumption_m2 = Column(Float)  # Расход краски на м²
+    print_price_m2 = Column(Float)  # Цена печати за м²
 
-    printer_rate_m2 = Column(Numeric(10, 2))  # Ставка печатника за м²
-    eyelet_rate = Column(Numeric(10, 2))  # Ставка за люверс
-    cutter_rate_m2 = Column(Numeric(10, 2))  # Ставка резчика за м²
-    payer_rate = Column(Numeric(10, 2))  # Ставка пайщика
+    printer_rate_m2 = Column(Float)  # Ставка печатника за м²
+    eyelet_rate = Column(Float)  # Ставка за люверс
+    cutter_rate_m2 = Column(Float)  # Ставка резчика за м²
+    payer_rate = Column(Float)  # Ставка пайщика
 
 Base.metadata.create_all(bind=engine)
